Add items related to deleting a user
Add the deleteTeacher() method to ru.dreremin.predefense.registration.sys
	.services.teachers.DeleteTeacherService class.
Add findByActorLogin() method to ru.dreremin.predefense.registration.sys
	.repositories.TeacherRepository interface.
Add the deleteAllByTeacherId() method to ru.dreremin.predefense
	.registration.sys.repositories.TeacherComissionRepository interface.
Add findByActorId() method to ru.dreremin.predefense.registration.sys
	.repositories.AdministratorRepository interface.
Add class ru.dreremin.predefense.registration.sys.services.admins
	.DeleteAdministratorService.
Add deleteTeacher(), deleteStudent(), deleteAdmin() methods to ru.dreremin
	.predefense.registration.sys.controllers.delete.DelenteUserController class.
Rename ru.dreremin.predefense.registration.sys.controllers.delete
	.DelentePersonController class to DelenteUserController.
Rename ru.dreremin.predefense.registration.sys.controllers.create
	.CreatePersonController class to CreateUserController.

+++ src/main/java/ru/dreremin/predefense/registration/sys/models/TeacherComission.java
@@ -20,14 +20,14 @@ public class TeacherComission {
 	private long id;
 	
 	@Column(name = "id_t_tcfk")
-	private long teacherId;
+	private int teacherId;
 	
 	@Column(name = "id_c_tcfk")
 	private int comissionId;
 	
 	public TeacherComission() {}
 	
-	public TeacherComission(long teacherId, int comissionId) {
+	public TeacherComission(int teacherId, int comissionId) {
 		this.teacherId = teacherId;
 		this.comissionId = comissionId;
 	}
